# docs/openapi.yaml
openapi: 3.0.3
info:
  title: "Siva Machine Works — Smart Manufacturing Platform API"
  version: "1.0.0"
  description: "Enterprise REST API for ERP, SCM, PDM/PLM, Dual-BOM, Production, QMS, Logistics, Finance, Analytics, and RAG AI Operations Assistant."
servers:
  - url: "http://localhost:8080/api/v1"
    description: Local Dev
  - url: "https://api.staging.sivamachineworks.com/api/v1"
    description: Staging
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: List Customers
      operationId: listCustomers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomers'
    post:
      tags: [Customers]
      summary: Create Customer
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
...
  /products:
    get:
      tags: [Products]
      summary: List Products
      operationId: listProducts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProducts'
    post:
      tags: [Products]
      summary: Create Product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{productId}:
    get:
      tags: [Products]
      summary: Get Product by ID
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/ProductIdParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Customer:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        region: { type: string }
    PaginatedCustomers:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Customer' } }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
    Product:
      type: object
      properties:
        id: { type: string }
        sku: { type: string, example: "SMW-HM-500" }
        name: { type: string }
        type:
          type: string
          enum: [RAW_MATERIAL, SUBASSEMBLY, FINISHED_GOOD]
    PaginatedProducts:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Product' } }
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
security:
  - BearerAuth: []